+++ build_runner/lib/build_runner.dart
@@ -9,3 +9,4 @@ export 'src/generate/build_result.dart';
 export 'src/generate/input_set.dart';
 export 'src/generate/phase.dart';
 export 'src/package_graph/package_graph.dart';
+export 'src/server/server.dart' show ServeHandler;